cmd/cqlmm: Implement the "create" command.
This is the command line for creating a new migration:

	$ cqlmm create MyNewMigration cql

You must specify "cql" at the end of the command, as this will enforce
intentional call signatures in the future, and allow for future support
of writing migrations in Go. Forcing the user to specify "cql" at the
end of the command (to specify the migration type) prevents any weird
"gotchas" from happening due to default behaviours.